all scripts intergrated and working together
